+++ ChangeLog.md
@@ -13,6 +13,7 @@
     - Define new '--check-album-path' and '--check-track-path' options for audio.pl list
     - TTP::commandExec() is extended to accept commands as object { command, args } to be able to bypass shell interpretation
     - audio.pl normalize is re-architectured to take advantage of new TTP::commandExec()
+    - audio.pl list has '--check-all-track' and '--check-all-album' options
 
 ### 4.19.0